# backend/services/api/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  domain.Project:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
    type: object
...
  handler.createProjectRequest:
    properties:
      description:
        type: string
      name:
        maxLength: 255
        minLength: 1
        type: string
    required:
    - name
    type: object
...
  handler.updateProjectRequest:
    properties:
      description:
        type: string
      name:
        maxLength: 255
        minLength: 1
        type: string
    required:
    - name
    type: object
...
  httputil.ErrorResponse:
    properties:
      details:
        type: string
      error:
        type: string
    type: object
...
  repo.MentionRow:
    properties:
      author:
        type: string
      brand_id:
        type: string
      created_at:
        type: string
      external_id:
        type: string
      id:
        type: string
      is_duplicate:
        type: boolean
      matched_keywords:
        items:
          type: string
        type: array
      matched_risk_words:
        items:
          type: string
        type: array
      project_id:
        type: string
      published_at:
        type: string
      sentiment_label:
        type: string
      sentiment_score:
        type: number
      source_id:
        type: string
      status:
        type: string
      text:
        type: string
      title:
        type: string
      url:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Brand reputation monitoring system API
  title: BrandRadar API
  version: "1.0"
paths:
...
  /projects:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.Project'
            type: array
      security:
      - BearerAuth: []
      summary: List user projects
      tags:
      - projects
    post:
      consumes:
      - application/json
      parameters:
      - description: Project data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.createProjectRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.Project'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/httputil.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a project
      tags:
      - projects
  /projects/{projectID}:
    delete:
      parameters:
      - description: Project ID
        in: path
        name: projectID
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httputil.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete project
      tags:
      - projects
    get:
      parameters:
      - description: Project ID
        in: path
        name: projectID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Project'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httputil.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get project by ID
      tags:
      - projects
    put:
      consumes:
      - application/json
      parameters:
      - description: Project ID
        in: path
        name: projectID
        required: true
        type: string
      - description: Project data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.updateProjectRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Project'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httputil.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update project
      tags:
      - projects
...
  /projects/{projectID}/feed:
    get:
      parameters:
      - description: Project ID
        in: path
        name: projectID
        required: true
        type: string
      - description: Filter by sentiment
        in: query
        name: sentiment
        type: string
      - description: Filter by status
        in: query
        name: status
        type: string
      - description: Filter since (RFC3339)
        in: query
        name: since
        type: string
      - description: Filter until (RFC3339)
        in: query
        name: until
        type: string
      - description: Filter by brand ID
        in: query
        name: brand_id
        type: string
      - description: Filter by source ID
        in: query
        name: source_id
        type: string
      - description: Text search
        in: query
        name: search
        type: string
      - description: Filter by risk words
        in: query
        name: has_risk
        type: boolean
      - default: 50
        description: Limit
        in: query
        name: limit
        type: integer
      - default: 0
        description: Offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/repo.MentionRow'
            type: array
      security:
      - BearerAuth: []
      summary: List mentions feed
      tags:
      - feed
  /projects/{projectID}/feed/{mentionID}:
    get:
      parameters:
      - description: Project ID
        in: path
        name: projectID
        required: true
        type: string
      - description: Mention ID
        in: path
        name: mentionID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/repo.MentionRow'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httputil.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get mention details
      tags:
      - feed
  /projects/{projectID}/feed/{mentionID}/dismiss:
    post:
      parameters:
      - description: Project ID
        in: path
        name: projectID
        required: true
        type: string
      - description: Mention ID
        in: path
        name: mentionID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Dismiss a mention
      tags:
      - feed
  /projects/{projectID}/feed/{mentionID}/duplicates:
    get:
      parameters:
      - description: Project ID
        in: path
        name: projectID
        required: true
        type: string
      - description: Mention ID
        in: path
        name: mentionID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/repo.MentionRow'
            type: array
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httputil.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List mention duplicates
      tags:
      - feed
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"